get initial review extent from review item (#1508)
refs #1506

The crux of the issue addressed is the map needs to be
zoomed to an editable extent before the UI can go into
review mode.  So get an initial review item (which now
supplies it's extent) and zoom the map.  Then instead of
loading a new review item we have a new method to
show the initial review item zoomed to.